updated order inventory and product details

+++ order-service/src/main/java/com/web/order/dtos/OrderItemResponse.java
@@ -2,12 +2,14 @@ package com.web.order.dtos;
 
 import com.web.order.enums.OrderStatus;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
         Long orderId,
         Long userId,
         Long productId,
         Integer quantity,
-        Double totalAmount,
+        BigDecimal totalAmount,
         OrderStatus orderStatus,
         String createdAt,
         String updatedAt
